+++ src/output/output_h264.cpp
@@ -195,7 +195,7 @@ namespace Mist{
         continue;
       }
       r["data"]["codecs"].append(codec);
-      r["data"]["tracks"].append(it->first);
+      r["data"]["tracks"].append((uint64_t)it->first);
       ++it;
     }
     webSock->sendFrame(r.toString());
@@ -305,7 +305,7 @@ namespace Mist{
       }
     }
     for (std::map<size_t, Comms::Users>::iterator it = userSelect.begin(); it != userSelect.end(); it++){
-      r["data"]["tracks"].append(it->first);
+      r["data"]["tracks"].append((uint64_t)it->first);
     }
     webSock->sendFrame(r.toString());
   }
@@ -448,7 +448,7 @@ namespace Mist{
         r["type"] = "info";
         r["data"]["msg"] = "Sending header";
         for (std::map<size_t, Comms::Users>::iterator it = userSelect.begin(); it != userSelect.end(); it++){
-          r["data"]["tracks"].append(it->first);
+          r["data"]["tracks"].append((uint64_t)it->first);
         }
         webSock->sendFrame(r.toString());
